修改

+++ BubbleSort.cpp
@@ -14,7 +14,7 @@ void BubbleSort::sort(int array[], int length)
 {
 	for (int j = length - 1; j > 0; j--)
 	{
-		for (size_t i = 0; i < j; i++)
+		for (int i = 0; i < j; i++)
 		{
 			if (array[i] > array[i + 1])
 			{
